Update plan.mdx for human-readable plan output
Uupdat plan.mdx for human-readable plan output to file.  This explains the process to output the terraform plan to a human-readable file that allows users to fully visualize and export the output of `terraform plan` which is currently machine-readable-only, which limits users' ability to read, understand and evaluate the outputs of `terraform plan`.